pyproject.toml
[project]
name = "utmn-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "fastapi>=0.135.1",
    "fastapi-cache2>=0.2.2",
    "langchain-gigachat>=0.5.0",
    "openai>=2.32.0",
    "orjson>=3.11.8",
    "passlib==1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-jose>=3.5.0",
    "redis==6.4.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
